add **prepare class** in util_functions and other module can Inheritance this class
